+++ packages/dso-toolkit/src/components/alert/alert.args.ts
@@ -3,6 +3,7 @@ import { ArgTypes } from "@storybook/types";
 import { HandlerFunction } from "@storybook/addon-actions";
 
 import { Alert, AlertStatus } from "./alert.models.js";
+import { noControl } from "../../storybook";
 
 export interface AlertArgs {
   status: AlertStatus;
@@ -10,12 +11,9 @@ export interface AlertArgs {
   compact: boolean;
   withRoleAlert: boolean;
   withButton: boolean;
+  closable: boolean;
+  dsoClose: HandlerFunction;
 }
-
-export const alertArgs: Pick<AlertArgs, "withButton" | "withRoleAlert"> = {
-  withButton: true,
-  withRoleAlert: false,
-};
 
 export const alertArgTypes: ArgTypes<AlertArgs> = {
   status: {
@@ -39,8 +37,17 @@ export const alertArgTypes: ArgTypes<AlertArgs> = {
       type: "boolean",
     },
   },
+  closable: {
+    control: {
+      type: "boolean",
+    },
+  },
   click: {
     action: "closed",
+  },
+  dsoClose: {
+    ...noControl,
+    action: "dsoClose",
   },
 };
 
@@ -54,5 +61,7 @@ export function alertArgsMapper<TemplateFnReturnType>(
     compact: a.compact,
     onClick: a.withButton ? () => a.click(a) : undefined,
     withRoleAlert: a.withRoleAlert,
+    closable: a.closable,
+    dsoClose: a.dsoClose,
   };
 }