+++ packages/dso-toolkit/src/components/alert/alert.models.ts
@@ -4,6 +4,12 @@ export interface Alert<TemplateFnReturnType> {
   status: AlertStatus;
   message: TemplateFnReturnType | string;
   compact?: boolean;
+  closable?: boolean;
+  dsoClose?: (e: CustomEvent<AlertCloseEvent>) => void;
   onClick?: (e: MouseEvent) => void;
   withRoleAlert?: boolean;
+}
+
+export interface AlertCloseEvent {
+  originalEvent?: MouseEvent | Event;
 }