+++ packages/react/src/components/alert/alert.react-template.tsx
@@ -3,14 +3,21 @@ import * as React from "react";
 
 import { DsoAlert } from "../../components";
 import { ComponentImplementation } from "../../templates";
+import { AlertCloseEvent, DsoAlertCustomEvent } from "@dso-toolkit/core";
 
 export const reactAlert: ComponentImplementation<Alert<JSX.Element>> = {
   component: "alert",
   implementation: "react",
   template: () =>
-    function alertTemplate({ message, onClick, status, compact, withRoleAlert }) {
+    function alertTemplate({ message, onClick, status, compact, withRoleAlert, closable, dsoClose }) {
       return (
-        <DsoAlert status={status} roleAlert={withRoleAlert} compact={compact}>
+        <DsoAlert
+          status={status}
+          roleAlert={withRoleAlert}
+          compact={compact}
+          closable={closable}
+          onDsoClose={(e: DsoAlertCustomEvent<AlertCloseEvent>) => dsoClose?.(e)}
+        >
           <div className="dso-rich-content">
             {typeof message === "string" ? <div dangerouslySetInnerHTML={{ __html: message }} /> : message}
             {onClick && (