+++ storybook/src/components/alert/alert.core-template.ts
@@ -3,14 +3,21 @@ import { html, nothing, TemplateResult } from "lit-html";
 import { unsafeHTML } from "lit-html/directives/unsafe-html.js";
 
 import { ComponentImplementation } from "../../templates";
+import { ifDefined } from "lit-html/directives/if-defined.js";
 
 export const coreAlert: ComponentImplementation<Alert<TemplateResult>> = {
   component: "alert",
   implementation: "core",
   template: ({ richContentTemplate }) =>
-    function alertTemplate({ status, message, compact, onClick, withRoleAlert }) {
+    function alertTemplate({ status, message, compact, onClick, withRoleAlert, closable, dsoClose }) {
       return html`
-        <dso-alert status=${status} ?role-alert=${withRoleAlert} ?compact=${compact}>
+        <dso-alert
+          status=${status}
+          ?role-alert=${withRoleAlert}
+          ?compact=${compact}
+          closable=${ifDefined(closable)}
+          @dsoClose=${dsoClose}
+        >
           ${richContentTemplate({
             children: html`
               ${typeof message === "string" ? unsafeHTML(message) : message}