+++ packages/react/src/components/list-button/list-button.stories.tsx
@@ -12,7 +12,7 @@ const meta: Meta<ListButtonArgs> = {
 
 export default meta;
 
-const { SingleSelect, MultiSelect, CssLegacyReadonlyCount } = listButtonStories({
+const { SingleSelect, MultiSelect } = listButtonStories({
   templateContainer,
   storyTemplates: (templates) => {
     const { listButtonTemplate } = templates;
@@ -23,4 +23,4 @@ const { SingleSelect, MultiSelect, CssLegacyReadonlyCount } = listButtonStories(
   },
 });
 
-export { SingleSelect, MultiSelect, CssLegacyReadonlyCount };
+export { SingleSelect, MultiSelect };