+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 ### Added
 * Icon: Mail ([#2918](https://github.com/dso-toolkit/dso-toolkit/issues/2918))
+* Alert: Nieuwe optie "closable" ([#2745](https://github.com/dso-toolkit/dso-toolkit/issues/2745))
 
 ## 👕 Release 67.3.2 - 2025-01-16