+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 ## Next
 
+## ⚖️ Release 68.0.0 - 2025-01-20
+
 ### Added
 * Icon: Mail ([#2918](https://github.com/dso-toolkit/dso-toolkit/issues/2918))
 * Alert: Nieuwe optie "closable" ([#2745](https://github.com/dso-toolkit/dso-toolkit/issues/2745))